Load article share component in article component
And also load share component for new articles loaded with AJAX
